% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check.thresholds}
\alias{check.thresholds}
\title{Check Thresholds}
\usage{
check.thresholds(this.row, vals.lst, lbls.lst, threshold.type,
  is.suitable = 1)
}
\description{
This function can be used to check EITHER an upper or a lower threshold.
}

