% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check.var.names}
\alias{check.var.names}
\title{Check variable names for spatialdemography compliance}
\usage{
check.var.names(var.names)
}
\arguments{
\item{var.names}{the list of variable names to be checked}
}
\description{
Check that each variable begins with a different first letter.
}

