% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compute.frag}
\alias{compute.frag}
\title{Create fragmentation data set for use}
\usage{
compute.frag(points, dataset, buffer.vec)
}
\arguments{
\item{points}{an input of points (even though the exporatories are plots, I'm using the center point)}

\item{dataset}{A data set to calculate fragmentation for}

\item{buffer.vec}{a vector of spatial scales at which to calculate fragmentation}
}
\description{
Right now, using CORINE, because it is easy and available, and I'm running out of time
Alternatives will be to follow up with Thomas Nauss about options or to contact Catrin Westfall
Catrin has digitized about a 2km buffer around each plot and classified the land use, so would
have much more precise fragmentation statistics.
(if that is really the case, why am I re-creating the wheel?
Because rumor has it she's not particularly interested in sharing the data.
NOTE: anything you set up for CORINE will also work with her data set, so set it up for CORINE
And then follow up with her - the comparison could make an interesting note.
I wonder if there is already an R tool for this?
NOTE: Fragmentation and heterogeneity are two words for effectively the same thing.
Or you could just try FRAGSTATS - that might be faster and easier.
Quick look doesn't look like there is an easy way to do what I want - just do what you know (even if you haven't done it in R.)
}

