% Generated by roxygen2 (4.0.1): do not edit by hand
\name{convert.sp.data.old}
\alias{convert.sp.data.old}
\title{Convert my.data to a format more easily used by the code}
\usage{
convert.sp.data.old(sp.data, model.file, year)
}
\description{
This version converted my.data to a SpeciesData.csv. Some of this code may be useful, but is not actually what I wanted in the end.
}

