% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do.basic.sdm}
\alias{do.basic.sdm}
\title{Basic SDM approach}
\usage{
do.basic.sdm(sp.col, sdm.data, col.index, training.proportion = 2/3)
}
\description{
Mainly for getting started - will want to do something more sophisticated for the final analysis
}

