% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do.descriptives}
\alias{do.descriptives}
\title{Do descriptives}
\usage{
do.descriptives(my.data, my.index, desc.path)
}
\arguments{
\item{my.data}{A data set with data for descriptive analysis}

\item{my.index}{an index coding data for different types of descriptive analysis}

\item{desc.path}{a folder to put the descriptive analysis results}
}
\description{
Basic descriptive analysis on the assembled dataset. Highlight key insights
}

