% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do.sdm.prefit}
\alias{do.sdm.prefit}
\title{SDM Evaluate the fit of an externally defined relationship}
\usage{
do.sdm.prefit(STUFF)
}
\description{
Here, all data is used as validation, as the suitability relationship
was defined based on an external data set
}

