% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do.sdm}
\alias{do.sdm}
\title{A main function to run the Species Distribution Model}
\usage{
do.sdm(sdm.data, my.index, col.index, sdm.type)
}
\arguments{
\item{my.data}{}

\item{my.index}{an index coding different column types.
type 0 corresponds to the ID field
type 1 corresponds to the species columns
type 2+ correspond to different suites of independent variables}

\item{col.index}{Contains indicators for whether or not to include the column as a predictor variable.}

\item{sdm.type}{1 = basic sdm (nothing fancy), 2 = uses biomod}
}
\description{
A main function to run the Species Distribution Model
}

