% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do.spatial.demography}
\alias{do.spatial.demography}
\title{Main function for generating the spatial demography outputs}
\usage{
do.spatial.demography(run.name, run.path, spfile, sp.locs, template.dir,
  file.ending, lpath, DisPath, opath, scale.vec, inputs.ending = "none")
}
\arguments{
\item{inputs.ending}{If the input files need a different ending than the output files, this allows that to happen.}
}
\description{
Main function for generating the spatial demography outputs
}

