% Generated by roxygen2 (4.0.1): do not edit by hand
\name{evaluate.spatialdemography}
\alias{evaluate.spatialdemography}
\title{Evaluate results of spatialdemography model runs}
\usage{
evaluate.spatialdemography(validation.file, model.file, validation.timestep,
  model.timestep, n.sp)
}
\arguments{
\item{validation.file}{A file containing the Validation data set}

\item{model.file}{A file containing the data set from model output}

\item{n.sp}{The number of species (probably could be extracted from the datasets)}
}
\description{
Goal is to compare output of spatialdemography model to a validation dataset
}

