% Generated by roxygen2 (4.0.1): do not edit by hand
\name{evaluate.spatialdemography.v2}
\alias{evaluate.spatialdemography.v2}
\title{Evaluate results of spatialdemography model runs using list objects as inputs.}
\usage{
evaluate.spatialdemography.v2(estimated.presences.lst, actual.presences.lst,
  n.sp)
}
\arguments{
\item{estimated.presences.lst}{A list with one entry for every species, giving predicted occupancy for each exploratory plot}

\item{actual.presences.lst}{A list with one entry for every species, giving observed occupancy for each exploratory plot}

\item{n.sp}{The number of species (equal to length of the two above lists!)}
}
\description{
Goal is to compare output of spatialdemography model to a validation dataset
}

