% Generated by roxygen2 (4.0.1): do not edit by hand
\name{explore.metacom}
\alias{explore.metacom}
\title{try out metacom package for examining Biodiversity Exploratories data}
\usage{
explore.metacom(bedata)
}
\description{
I really should try to get to the point where I am talking with the plant people more regularly
this approach may be leading to some serious redundancies.
Should have taken better advantage of the exploratories assemblies. But if it was in APRIL,
I'd have been in a position to say something cool!
}

