% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract.cell.values}
\alias{extract.cell.values}
\title{Extract the value of a single cell from the SpeciesData.csv}
\usage{
extract.cell.values(predicted.presences.lst, output.file, focal.cell.position,
  validation.timestep)
}
\description{
Extract the value of a single cell from the SpeciesData.csv
}

