% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extrapolation.test}
\alias{extrapolation.test}
\title{Extrapolation test}
\usage{
extrapolation.test(eval.dat, train.dat)
}
\arguments{
\item{eval.dat}{the data to be evaluated for whether they fall outside the predicted range}

\item{train.dat}{the data set used in training}
}
\description{
Test whether examined values fall outside the range of the training data set
NOTE: AT PRESENT THIS DOES NOT TEST WHETHER *COMBINIATIONS* fall outside the range of the training data set
But that probably could be done with a convex hull or with the UTC approach
}

