% Generated by roxygen2 (4.0.1): do not edit by hand
\name{frag.pca}
\alias{frag.pca}
\title{Examine fragmentation data and consider variable reduction}
\usage{
frag.pca(frag.dat, descriptive.analysis = 0)
}
\value{
A list containing the first two grassland principle components, the means used for centering
 and the amount of variation explained by the two principle components.
}
\description{
This function was used to explore the fragmentation data. Based on the exploration
 I decided to compute principle component scores separately for each variable type
 i.e. just for grassland scales or just for forest scales.
 For my question, I have decided that I will start with just the grassland data;
 if people disagree, I can add the other percent cover types.
}
\details{
A message will tell how much variation is explained in these variables, but will not output it
 But I should think about extracting the data as a list, with that information as part of the list
 object.
}

