% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.index}
\alias{get.index}
\title{Get Index}
\usage{
get.index(data.lst)
}
\arguments{
\item{A}{list of data sets requiring the index.
Each data set must have one, and only one identical ID field.}
}
\description{
Provides an index to which columns come from which data source for later data manipulation
Note that the first column in the combined data set needs to be plotyear, the ID field
Also note that the species data must be first in data.lst
}

