% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.survival}
\alias{get.survival}
\title{Read in survival data}
\usage{
get.survival(survival.file, ludata)
}
\arguments{
\item{survival.data}{The survival data from Helge, Ute, & Eva}

\item{ludata}{Land use data from the Biodiversity Exploratories}
}
\description{
Also relevant as response traits to the metacommunity model
Data from Helge, Ute, & Eva
}
\details{
Hmm, I've been using plotyear as a joining method, but these data aren't associated with a particular year.
Need to add a second join to the main code to add the datasets that are not year specific
}

