% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.thresholds.old}
\alias{get.thresholds.old}
\title{Get thresholds}
\usage{
get.thresholds.old(train.dat, sp.col, col.index, sp.name, fs.presences)
}
\description{
Calculate a simple bioclimatic envelope based on extremes of known occurrence
}

