% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.thresholds}
\alias{get.thresholds}
\title{Get thresholds}
\usage{
get.thresholds(train.dat, sp.col, col.index, sp.name, fs.presences)
}
\description{
Calculate a simple bioclimatic envelope based on extremes of known occurrence.\
Changed from previous version in that only extremes from presences are used as thresholds.
This is a dubious approach at best.
}

