% Generated by roxygen2 (4.0.1): do not edit by hand
\name{merge.data}
\alias{merge.data}
\title{Merge data}
\usage{
\method{merge}{data}(data.lst, id.field)
}
\arguments{
\item{data.lst}{a list of data files to be merged}

\item{id.field}{the field to use for merging the data}
}
\description{
Function to handle merging of data sets for species distribution modeling
}

