% Generated by roxygen2 (4.0.1): do not edit by hand
\name{metacom.analysis}
\alias{metacom.analysis}
\title{Use metacom package to assess the biotic filter for Biodiversity Exploratories data}
\usage{
metacom.analysis(sp.data, year)
}
\arguments{
\item{sp.data}{Species data with a plotyear column containing plots and years}

\item{year}{the year of data to subset to. (what happens if you do it across years instead of across plots?)}
}
\description{
#**# Which data does it take as an input? Need to remember how I was doing this.
NOTE: requires an updated version of vegan to work.
}

