% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prepare.sp.locations.old}
\alias{prepare.sp.locations.old}
\title{Create species locations file}
\usage{
prepare.sp.locations.old(sp.locs, my.data, start.year, col.index)
}
\description{
This approach was designed to convert biodiversity exploratories data into
the spatialdemography model. Problem is, plots are not adjacent to one another
so that really isn't an appropriate analysis technique for this dataset.
}

