% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prepare.sp.locations}
\alias{prepare.sp.locations}
\title{Create species locations file}
\usage{
prepare.sp.locations(sp.locs, sp.plot.dat, pool.type, n.cells)
}
\description{
Given a landscape, create initial species locations.
}
\details{
Note to self: think about whether you want species initialization to depend on landscape cells
I.e. to what extent are species drawn randomly, and to what extent are they drawn based on the
environmental conditions present? (we could test some interesting hypotheses with this model
setup.)
}

