% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prepare.spfile}
\alias{prepare.spfile}
\title{Prepare species file}
\usage{
prepare.spfile(spbase, spfile, thresholds)
}
\arguments{
\item{spbase}{The base file to append response traits to}

\item{spfile}{The new species file to be created}

\item{thresholds}{Additional thresholds to be appended to the species file}
}
\description{
Add additional response traits to an existing species file containing base
vital rates, dispersal parameters, and (any) known response traits.
}
\details{
thresholds should be a set of nested lists.
The first list should have one entry for each species
Each species list entry should be a list of 3 vectors
vector 1 gives variable names, vector 2 lower thresholds, vector 3 = upper thresholds.
}

