% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.climate}
\alias{read.climate}
\title{Read in (yearly) climate data}
\usage{
read.climate(inpath)
}
\description{
Read in climate data that has been quality controlled.
Note that 2008 data should probably be dropped:
it looks like some further quality control is needed.
}

