% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.frag}
\alias{read.frag}
\title{Read in fragmentation data created by make.frag function}
\usage{
read.frag(pc.file)
}
\arguments{
\item{pc.file}{File containing the results of a principle component analysis used to reduce multicollinearity.}
}
\description{
Read in fragmentation data created by make.frag function
}

