% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.sp.data}
\alias{read.sp.data}
\title{Read in be_data file}
\usage{
read.sp.data(inpath, sp.lookup.file, dropcol = TRUE, plotyear.rows = TRUE,
  write.data = T,
  out.file = "C:/docs/beplants/datasets/plants/PlantDataBE_rf.csv")
}
\arguments{
\item{inpath}{the input path for the data}

\item{dropcol}{Whether or not to drop the added empty columns to make the landscape square for spatialdemography}
}
\description{
Reads the species data file
Reformat the file so that species are columns and plot-years are rows.
NOTE: species are coded by number #**# NEED TO LOOK UP FILE THAT HAS species ID's
NOTE: Not every species has an entry for every year!
}

