% Generated by roxygen2 (4.0.1): do not edit by hand
\name{reformat.sb.plant.data}
\alias{reformat.sb.plant.data}
\title{Reformat steffen's plant data}
\usage{
reformat.sb.plant.data(in.data, out.path)
}
\description{
R code to put the plant data from Steffen Boch into the format needed by SpatialDemography
NOTE: Prior to this code, the .csv was opened in Excel, split based on a ";" delimiter, then saved with a "," delimiter.
Note: function never run as a function, but was run using the commented out text at the beginning.
}

