% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Reimagine}
\alias{Reimagine}
\title{Function to fix problems with Imagine function in metacom package}
\usage{
Reimagine(comm, col = c(0, 1), ordinate = TRUE, scores = 1, fill = TRUE,
  xlab = "Species", ylab = "Site", yline = 2, xline = 2,
  sitenames = rownames(comm), speciesnames = colnames(comm),
  binary = TRUE, cex.x = 1.5, cex.y = 1.5, staggered.x.labels = FALSE,
  staggered.y.labels = FALSE)
}
\description{
Copied the Imagine function from metacom, and then adapted it to my needs
First real use of the Open Source part of open source!
Well, I've got it plotting my labels, but they're so tiny, that they're illegible!
I think I need staggered labels = have some above and some below
}

