% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setup.predictors}
\alias{setup.predictors}
\title{Main function for reading in x variables}
\usage{
setup.predictors(sp.data, predictors.vec, predictors.paths, start.year = 2008,
  end.year = 2013)
}
\description{
Main function for reading in x variables
}

