% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spdem.extract}
\alias{spdem.extract}
\title{Extract data for a species distribution model from spatialdemography model}
\usage{
spdem.extract(data.file, timestep = "all", adults.only = TRUE)
}
\arguments{
\item{data.file}{This input should be in the format of a species locations file from spatialdemography, or in the format of the SpeciesData.csv file created as a model output}

\item{timestep}{Will allow a user to restrict the data extraction to a particular timestep.}
}
\description{
I think I wrote a function doing exactly this within spatialdemography, to produce species locations files?
Look to see if this is redundant code!
}

