% Generated by roxygen2 (4.0.1): do not edit by hand
\name{subset.by.traits}
\alias{subset.by.traits}
\title{Subset by traits}
\usage{
\method{subset}{by.traits}(sp.data, trait.data, link, trait.lst, trait.vals.lst)
}
\description{
Plan is to give an option to subset the dataset based on preselected traits
}

