% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summarize.validation}
\alias{summarize.validation}
\title{Create a summary from multiple simulations}
\usage{
summarize.validation(sim.lst, n.sp)
}
\description{
Think about whether there is a better way to code this!
Add options for median & quantiles later
}

