% Generated by roxygen2 (4.0.1): do not edit by hand
\name{try.biomod}
\alias{try.biomod}
\title{Test biomod2 with Biodiversity Exploratories data}
\usage{
try.biomod(sp.name, fs.presences.all, env.lyrs)
}
\arguments{
\item{sp.name}{The name of the focal species}

\item{fs.presences.all}{List of presences and absences of the species}

\item{env.lyrs}{a list of environmental layers to be examined.}
}
\description{
NOTE: some things are being plotted at odd times and in
odd ways. This needs improvement
}

