% Generated by roxygen2 (4.0.1): do not edit by hand
\name{unfinished.function}
\alias{unfinished.function}
\title{I switched to processing in ArcGIS, but there are already fragmentation data from Catrin Westphal.
Calculate Fragmentation Data}
\usage{
unfinished.function(NOTHING)
}
\description{
Turned out this was hard to do in R, so I did it in QGIS instead. Here are the steps I took:
}
\details{
Here are the steps I took:
\tabular{ll}{
Download CORINE 2006 (COR06) Data   \tab Accepted continent-wide fragmentation data \cr
Cut COR06 data to match each exploratory \tab work with smaller data files, done earlier in ArcGIS \cr
Reproject COR06 to match exploratories (UTM 32 for HAI & ALB, 33 for SCH). Chose cubic resampling - oh, I bet I know where I got the all forest cells from last time... \tab Better to have everything in same coord. system
Start with plot shapefiles (UTM 32 for HAI & ALB, 33 for SCH)  \tab I made these with ArcGIS based on the .kml files \cr
Set QGIS to do on-the-fly converstions \tab Helps get layers to line up \cr
Buffer plot shapefiles (250 m, 1000 m, 5000 m)      \tab \cr
}
}

