% Generated by roxygen2 (4.0.1): do not edit by hand
\name{validate.sdm}
\alias{validate.sdm}
\title{Validate SDM}
\usage{
validate.sdm(fs.presences.validation, suitability.vec)
}
\description{
Compare the predicted suitability from the classification with the observed
distribution in the validation data set
}

