% Generated by roxygen2 (4.0.1): do not edit by hand
\name{visualize.sdm}
\alias{visualize.sdm}
\title{Create a visual of the SDM predictions}
\usage{
visualize.sdm(sp.name, sp.image, validation.diagnostics, plotlabels,
  suitability.vec.all, fs.presences.all, training.index, extrapolation.index)
}
\arguments{
\item{plotlabels}{A list of the names of each plot, for use in identifying particular plots}

\item{suitability.vec.all}{}

\item{training.index}{Index to indicate which cells were in the training (T) and which were in the validation sets (F).}
}
\description{
For visually assessing the quality of the results. Plan is to plot suitability as the color of the cell,
plot a symbol to indicate presence or absence, color the cell border to indicate training or validation data
and finally to have a second cell border that indicates whether or not the cell contains extrapolated values beyond the validation set
The approach to changing the background color (later adapted to producing additional rectangles) was from Joran and stackoverflow.
Originally from an R help thread.
}

