% Generated by roxygen2 (4.0.1): do not edit by hand
\name{year.subset}
\alias{year.subset}
\title{Year subset}
\usage{
year.subset(in.data, year)
}
\arguments{
\item{in.data}{Data set (e.g. sp.data) with a plotyear column containing plots and years and the rest as data columns}

\item{year}{the year of data to subset to. (what happens if you do it across years instead of across plots?)}
}
\description{
Restrict to one year of the exploratories data, convert plots to row.names and drop plotyear column.
May want to back update other parts of the code that do essentially this!
}

