% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.survival}
\alias{get.survival}
\title{Read in survival data}
\usage{
get.survival(survival.file, ludata)
}
\arguments{
\item{survival.data}{The survival data from Helge, Ute, & Eva}

\item{ludata}{Land use data from the Biodiversity Exploratories}
}
\description{
Convert survival data from Helge, Eva, & Ute into response traits for
use with the spatialdemography model
}
\details{
Here, I read in the survival data and the land use data. For now,
we just use survival at the 7th time point and LUI averaged from 2012 and 2013
(but based on a global LUI). I averaged survival for all individuals within
a plot to avoid pseudo-replication, and then did a linear regression of
survival probability vs. LUI.
The parameter estimates and regressions are then returned from this function
for use in the spatialdemography model.
}

