% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make.frag}
\alias{make.frag}
\title{Read in fragmentation data}
\usage{
make.frag(frag.file, start.year, end.year, pc.file)
}
\description{
Using Catrin Westphal & colleagues data
Note: I saved the exploratories data as a .csv, to make it easier for me to import (because I don't know how to do tab delimiters in R, apparently)
}

